Fix warnings, working on x64

+++ src/windows/regex/engine.c
@@ -130,7 +130,7 @@ matcher(struct re_guts *g, char *string, size_t nmatch, regmatch_t pmatch[],
     int eflags)
 {
 	char *endp;
-	int i;
+	size_t i;
 	struct match mv;
 	struct match *m = &mv;
 	char *dp;
@@ -974,7 +974,7 @@ print(struct match *m, char *caption, states st, int ch, FILE *d)
 	(void)fprintf(d, "\n");
 }
 
-/* 
+/*
  - at - print current situation
  */
 static void
